Add static analysis and memory checking infrastructure
Implement comprehensive code quality tools:
- Enhanced compiler warnings (18+ additional warning flags)
- Cppcheck static analysis integration
- Valgrind memory leak detection

Compiler Warnings:
- Add -Wshadow, -Wconversion, -Wsign-conversion
- Add -Wold-style-cast, -Wdouble-promotion
- Add -Wnull-dereference, -Wimplicit-fallthrough
- Add GCC-specific: -Wduplicated-cond, -Wlogical-op

Static Analysis:
- Cppcheck target for warning/performance/portability checks
- Suppression of style-only issues
- Focus on real bugs and performance problems

Memory Checking:
- Valgrind integration with full leak checking
- Separate targets for silent run and detailed reports
- Combined 'analyze' target for all checks

Code Improvements:
- Fix all sign conversion warnings (int → size_t)
- Replace old-style casts with static_cast/reinterpret_cast
- Eliminate redundant EAGAIN/EWOULDBLOCK checks
- Use resize() instead of substr() for performance
- Return strings by const reference where appropriate

Results:
- Zero compiler warnings with enhanced flags ✓
- Cppcheck passes cleanly ✓
- Valgrind: 0 memory leaks, 0 errors ✓

Build Targets:
- make cppcheck    - Run static analysis
- make valgrind    - Run memory leak check
- make analyze     - Run all checks

+++ src/command_interpreter.cpp
@@ -40,8 +40,8 @@ TestCommand CommandInterpreter::interpret(const std::map<std::string, std::strin
             break;
 
         case BehaviorType::CLOSE_AFTER_PARTIAL:
-            cmd.bytes_before_close = parseInteger(
-                query_params.count("bytes") ? query_params.at("bytes") : "", 0);
+            cmd.bytes_before_close = static_cast<size_t>(parseInteger(
+                query_params.count("bytes") ? query_params.at("bytes") : "", 0));
             break;
 
         case BehaviorType::SLOW_RESPONSE:
